+++ data/Pickle.txt
@@ -1,3 +1,4 @@
 D | 0 | movie | 2020-04-04 1800
 E | 0 | movie | 2020-04-04 1800 | 2020-04-04 1900
 D | 0 | read book | 2019-12-12 1800
+T | 1 | find people
