+++ src/main/java/pickle/exception/PickleException.java
@@ -1,3 +1,5 @@
+package pickle.exception;
+
 public class PickleException extends Exception{
 
     public PickleException(String description) {
