+++ src/main/java/pickle/task/Deadline.java
@@ -1,3 +1,5 @@
+package pickle.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
