+++ src/main/java/pickle/task/Event.java
@@ -1,4 +1,5 @@
-import java.time.LocalDate;
+package pickle.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
