+++ src/main/java/pickle/task/Task.java
@@ -1,3 +1,5 @@
+package pickle.task;
+
 public abstract class Task {
     private String description;
     private boolean status;
